Fixed crash with iOS11.
For some reason, constraints for the segmented control that were perfectly fine before iOS11, now crash at runtime.
The good news is that said constraints are not needed anymore to center the segmented control in the toolbar, because somehow the system does it automatically.
Worked around the issue by adding the constraints only for iOS < 11.